Allow "conw" to be used even if conw_on=false
Temporary fix to allow "conw" to be used even if "conw auto" toggles conw off, or if the user enters "conw off" to toggle conw off.

This will temporarily allow conw to be used in these cases while we work on separating the idea of "auto conw" away from turning conw on/off.